StockNotifier/requirements.txt
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
awscli==1.38.11